% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalib-package.R
\docType{package}
\name{datalib-package}
\alias{datalib}
\alias{datalib-package}
\title{datalib: Manage Versioned Survey Microdata Libraries}
\description{
Client for a standardized, versioned survey microdata library (masters and harmonized adaptations) organized as <root>/<CCC>/<CCC>_<YYYY>_<SSSS>/<vintage>/Data/{Original,Stata,R}/. Enumerates countries, surveys, vintages and adaptations; resolves paths; loads and merges survey modules from Stata .dta files. Part of a trilingual suite also available for 'Stata' and 'Python'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/unicef-drp/datalib-unicef}
  \item Report bugs at \url{https://github.com/unicef-drp/datalib-unicef/issues}
}

}
\author{
\strong{Maintainer}: Joao Pedro Azevedo \email{jpazevedo@unicef.org}

}
\keyword{internal}
